Vision tokenizer moved to apertus.py and apertus.py multimodal application cleaned
